changed dec to dbl

+++ Sud/Models/ViewModels/ShoppingCartViewModel.cs
@@ -8,6 +8,6 @@ namespace Sud.Models.ViewModels
     public class ShoppingCartViewModel
     {
         public ShoppingCart ShoppingCart { get; set; }
-        public decimal ShoppingCartTotal { get; set; }
+        public double ShoppingCartTotal { get; set; }
     }
 }
